+++ test/acceptance/export/kml.js
@@ -30,6 +30,38 @@ var checkDecimals = function(x, dec_sep){
         return 0;
 }
 
+// Check if an attribute is in the KML output
+//
+// NOTE: "name" and "description" attributes are threated specially
+//       in that they are matched in case-insensitive way
+//
+var hasAttribute = function(kml, att) {
+
+  // Strip namespace:
+  //https://github.com/polotek/libxmljs/issues/212
+  kml = kml.replace(/ xmlns=[^>]*>/, '>');
+
+  var doc = libxmljs.parseXmlString(kml);
+  //console.log("doc: " + doc);
+  var xpath;
+
+  xpath = "//SimpleField[@name='" + att + "']";
+  if ( doc.get(xpath) ) return true;
+
+  xpath = "//Placemark/" + att;
+  if ( doc.get(xpath) ) return true;
+
+  var lcatt = att.toLowerCase();
+  if ( lcatt == 'name' || lcatt == 'description' ) {
+    xpath = "//Placemark/" + lcatt;
+    if ( doc.get(xpath) ) return true;
+  }
+
+  //if ( lowerkml.indexOf('simplefield name="'+ loweratt + '"') != -1 ) return true;
+  //if ( lowerkml.indexOf('<'+loweratt+'>') != -1 ) return true;
+  return false;
+}
+
 // KML tests
 
 test('KML format, unauthenticated', function(done){
@@ -46,9 +78,9 @@ test('KML format, unauthenticated', function(done){
         var checkfields = {'Name':1, 'address':1, 'cartodb_id':1, 'the_geom':0, 'the_geom_webmercator':0};
         for ( var f in checkfields ) {
           if ( checkfields[f] ) {
-            assert.ok(row0.indexOf('SimpleData name="'+ f + '"') != -1, "result does not include '" + f + "'");
+            assert.ok(hasAttribute(row0, f), "result does not include '" + f + "': " + row0);
           } else {
-            assert.ok(row0.indexOf('SimpleData name="'+ f + '"') == -1, "result includes '" + f + "'");
+            assert.ok(!hasAttribute(row0, f), "result includes '" + f + "'");
           }
         }
         done();
@@ -103,9 +135,9 @@ test('KML format, skipfields', function(done){
         var checkfields = {'Name':1, 'address':0, 'cartodb_id':0, 'the_geom':0, 'the_geom_webmercator':0};
         for ( var f in checkfields ) {
           if ( checkfields[f] ) {
-            assert.ok(row0.indexOf('SimpleData name="'+ f + '"') != -1, "result does not include '" + f + "'");
+            assert.ok(hasAttribute(row0, f), "result does not include '" + f + "': " + row0);
           } else {
-            assert.ok(row0.indexOf('SimpleData name="'+ f + '"') == -1, "result includes '" + f + "'");
+            assert.ok(!hasAttribute(row0, f), "result includes '" + f + "'");
           }
         }
         done();
@@ -190,8 +222,8 @@ test('GET /api/v1/sql as kml with no rows', function(done){
         method: 'GET'
     },{ }, function(res){
         assert.equal(res.statusCode, 200, res.body);
-        var body = '<?xml version="1.0" encoding="utf-8" ?>\n<kml xmlns="http://www.opengis.net/kml/2.2">\n<Document><Folder><name>sql_statement</name>\n</Folder></Document></kml>\n';
-        assert.equal(res.body, body);
+        var body = '<?xml version="1.0" encoding="utf-8" ?><kml xmlns="http://www.opengis.net/kml/2.2"><Document><Folder><name>sql_statement</name></Folder></Document></kml>';
+        assert.equal(res.body.replace(/\n/g,''), body);
         done();
     });
 });
@@ -207,8 +239,8 @@ test('GET /api/v1/sql as kml with ending semicolon', function(done){
         method: 'GET'
     },{ }, function(res){
         assert.equal(res.statusCode, 200, res.body);
-        var body = '<?xml version="1.0" encoding="utf-8" ?>\n<kml xmlns="http://www.opengis.net/kml/2.2">\n<Document><Folder><name>sql_statement</name>\n</Folder></Document></kml>\n';
-        assert.equal(res.body, body);
+        var body = '<?xml version="1.0" encoding="utf-8" ?><kml xmlns="http://www.opengis.net/kml/2.2"><Document><Folder><name>sql_statement</name></Folder></Document></kml>';
+        assert.equal(res.body.replace(/\n/g,''), body);
         done();
     });
 });